refactor and add kfp helpers to express helpers
