+++ examples/pipelines/components.config
