+++ examples/pipelines/finetune_stable_diffusion/components/clip_retrieval_component/build_image.sh
@@ -14,7 +14,7 @@
 # limitations under the License.
 
 # Read component configs
-. ../../components.config
+. ../../../components.config
 
 # Set some variables
 ARTIFACT_PATH=${PROJECT_ARTIFACT_PATH}
