--- examples/pipelines/finetune_stable_diffusion/pipelines/requirements.txt
@@ -1,3 +1,0 @@
-google-api-core==2.8.1
-kfp==1.8.14
-kubernetes==18.20.0
