+++ examples/pipelines/finetune_stable_diffusion/pipelines_config/sd_finetuning_config.py
