+++ examples/pipelines/finetune_stable_diffusion/sd_finetuning_pipeline.py
@@ -2,21 +2,25 @@
 using clip retrieval on the LAION dataset"""
 # pylint: disable=import-error
 import logging
+import sys
+import os
 from typing import Optional
 
 from kfp import components as comp
 from kfp.v2 import dsl
 from kubernetes import client as k8s_client
 
-from config.general_config import GeneralConfig, KubeflowConfig
-from config.sd_finetuning_config import StableDiffusionFinetuningConfig as SDConfig
-from helpers.upload import compile_and_upload_pipeline
+sys.path.insert(0, os.path.abspath('..'))
+
+from config import GeneralConfig, KubeflowConfig
+from pipelines_config.sd_finetuning_config import StableDiffusionFinetuningConfig as SDConfig
+from express.kfp_utils import compile_and_upload_pipeline
 
 logging.basicConfig(level=logging.INFO)
 LOGGER = logging.getLogger(__name__)
 
 sd_finetuning_component = comp.load_component(
-    '../components/sd_finetuning_component/component.yaml')
+    'components/sd_finetuning_component/component.yaml')
 
 
 # Pipeline
