+++ pyproject.toml
@@ -40,9 +40,12 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = "^3.8"
-dataclasses-json = "^0.5.7"
-pandas = "^1.3.5"
 datasets = "2.10.1"
+dataclasses-json = "^0.5.7"
+google-api-core = "^2.8.1"
+kfp = "^1.8.14"
+kubernetes= "^18.20.0"
+pandas = "^1.3.5"
 
 [tool.poetry.group.test.dependencies]
 liccheck = "^0.7.3"
